Optimize environment access for transitive functions
In the current solution, transitive functions used an additional function parameter environment that contained the variables accessed by the transitive function from the outer scope. This approach has the benefit to be stateless, however, requires that this additional argument is added for every call. As its not possible to identify all call sites of a transitive function without a very sophisticated data flow analysis, a wrapper has been introduced that adds the additional environment variable (fix function declaration instead of identifing the call sites). This requires that the arguments for every call are copied into a new array into which also the environment variable is pushed what has a significant --- negative --- impact on performance. Besides, it also may break functions that uses arguments to identify if an optional argument is present or not.

The new approach uses module-globals for the variables used inside of parallel-functions from the outer scope. These variables are initialized in the entry function (the one that is called from parallel) and reset to undefined when the function completed (successful or not). This requires less code rewriting and has a significant better performance than the existing approach (mandelbrot computed in 24s instead of 33s)