Emit Error for undeclared Variables
Emit an error for undefined variables accessed inside parallel functors. This ensures that it's impossible that one function access environment variables of another function.
This is needed as the global environment variables change the semantics if a function access a undeclared variable that is part of the environment from another functor. Without rewriting this would result in a ReferenceError that the variable is not declared. But with the extracted environment this is no longer the case. Therefore check it at compile time.